+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -391,6 +391,7 @@ module ActiveRecord
         UNIQUE_VIOLATION      = "23505"
         SERIALIZATION_FAILURE = "40001"
         DEADLOCK_DETECTED     = "40P01"
+        LOCK_NOT_AVAILABLE    = "55P03"
 
         def translate_exception(exception, message)
           return exception unless exception.respond_to?(:result)
@@ -410,6 +411,8 @@ module ActiveRecord
             SerializationFailure.new(message)
           when DEADLOCK_DETECTED
             Deadlocked.new(message)
+          when LOCK_NOT_AVAILABLE
+            TransactionTimeout.new(message)
           else
             super
           end